Check for symbol before comparing with flow-storm/bookmark
There is a weird issue when tracing a compilation of the ClojureScript
compiler where just doing (= expr-val 'flow-storm/bookmark), when
expr-val is a lazy-seq it forces it to realize, which is wrong.
This doesn't happen with all LazySeqs so not sure what is going on.
Let's check for symbol before checking with equality.

...
[clojure.lang.LazySeq force "LazySeq.java" 50]
[clojure.lang.LazySeq realize "LazySeq.java" 89]
[clojure.lang.LazySeq seq "LazySeq.java" 106]
[clojure.lang.LazySeq equiv "LazySeq.java" 147]
[clojure.lang.Util pcequiv "Util.java" 124]
[clojure.lang.Util equiv "Util.java" 32]
[flow_storm.runtime.indexes.api$add_expr_exec_trace invokeStatic "api.cljc" 331]
...